pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
[project]
name = "mcp-immo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.2.0",
    "huggingface-hub>=1.9.0",
    "joblib>=1.5.3",
    "langchain>=1.2.15",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-mistralai>=1.1.2",
    "langgraph>=1.1.6",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
]
